feat: add formatter configuration, package files, and GitHub workflows for Solidity code formatting

+++ src/libraries/DepositDataRootGenerator.sol
@@ -2,39 +2,18 @@
 pragma solidity ^0.8.13;
 
 library depositDataRootGenerator {
-    uint constant GWEI = 1e9;
+    uint256 constant GWEI = 1e9;
 
-    function generateDepositDataRoot(
-        bytes calldata pubkey,
-        bytes calldata signature,
-        bytes calldata withdrawal_credentials,
-        uint256 _amountIn
-    ) public pure returns (bytes32) {
-
-        uint deposit_amount = _amountIn / GWEI;
+    function generateDepositDataRoot(bytes calldata pubkey, bytes calldata signature, bytes calldata withdrawal_credentials, uint256 _amountIn) public pure returns (bytes32) {
+        uint256 deposit_amount = _amountIn / GWEI;
         bytes memory amount = to_little_endian_64(uint64(deposit_amount));
 
         bytes32 pubkey_root = sha256(abi.encodePacked(pubkey, bytes16(0)));
-        bytes32 signature_root = sha256(
-            abi.encodePacked(
-                sha256(abi.encodePacked(signature[:64])),
-                sha256(abi.encodePacked(signature[64:], bytes32(0)))
-            )
-        );
-        return
-            sha256(
-                abi.encodePacked(
-                    sha256(
-                        abi.encodePacked(pubkey_root, withdrawal_credentials)
-                    ),
-                    sha256(abi.encodePacked(amount, bytes24(0), signature_root))
-                )
-            );
+        bytes32 signature_root = sha256(abi.encodePacked(sha256(abi.encodePacked(signature[:64])), sha256(abi.encodePacked(signature[64:], bytes32(0)))));
+        return sha256(abi.encodePacked(sha256(abi.encodePacked(pubkey_root, withdrawal_credentials)), sha256(abi.encodePacked(amount, bytes24(0), signature_root))));
     }
 
-    function to_little_endian_64(
-        uint64 value
-    ) internal pure returns (bytes memory ret) {
+    function to_little_endian_64(uint64 value) internal pure returns (bytes memory ret) {
         ret = new bytes(8);
         bytes8 bytesValue = bytes8(value);
         // Byteswapping during copying to bytes.
